LayoutTests/fast/dom/javascript-url-exception-isolation-expected.txt

CONSOLE MESSAGE: 42
CONSOLE MESSAGE: line 1: SyntaxError: Unexpected token '<'
Exceptions thrown in javascript URLs should not propagate to the main script.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS caughtException is false
PASS caughtException is false
PASS successfullyParsed is true

TEST COMPLETE
